+++ src/views/Visualize.vue
@@ -88,6 +88,10 @@
         geocoding: null,
         lastUpdate: null,
 
+        mapCenterLatitude: process.env.VUE_APP_VISU_MAP_CENTER_LATITUDE,
+        mapCenterLongitude: process.env.VUE_APP_VISU_MAP_CENTER_LONGITUDE,
+        mapCenterZoomLevel: process.env.VUE_APP_VISU_MAP_ZOOM_LEVEL,
+
         dateFilter: _today.toISOString().split('T')[0],
         allowedDates: [],
         datePickerFormat: {
@@ -175,6 +179,30 @@
         this.error = error;
       }
 
+      const reportDataString = localStorage.getItem('report-data');
+      if (reportDataString) {
+
+        const reportData = JSON.parse(reportDataString);
+        try {
+          this.mapCenterLatitude = this.geocoding[reportData.postalCode].latitude;
+          this.mapCenterLongitude = this.geocoding[reportData.postalCode].longitude;
+
+          if (process.env.VUE_APP_REVERSE_GEO_LAT_LNG) {
+            this.mapCenterLatitude = this.geocoding[reportData.postalCode].longitude;
+            this.mapCenterLongitude = this.geocoding[reportData.postalCode].latitude;
+          }
+
+          if (process.env.VUE_APP_VISU_MAP_ZOOM_LEVEL_LOCAL) {
+            this.mapCenterZoomLevel = process.env.VUE_APP_VISU_MAP_ZOOM_LEVEL_LOCAL;
+          } else {
+            this.mapCenterZoomLevel = +process.env.VUE_APP_VISU_MAP_ZOOM_LEVEL + 2;
+          }
+        } catch (error) {
+
+        }
+
+      }
+
       try {
         await this.loadLeaflet();
       } catch (error) {
@@ -234,6 +262,10 @@
 
                 if (location.longitude && location.latitude) {
                   location.coordinates = [+location.latitude, +location.longitude];
+
+                  if (process.env.VUE_APP_REVERSE_GEO_LAT_LNG) {
+                    location.coordinates = [+location.longitude, +location.latitude];
+                  }
                 } else {
                   continue;
                 }
@@ -308,12 +340,16 @@
       },
       loadLeaflet: async function () {
 
+        console.log(this.mapCenterLatitude);
+        console.log(this.mapCenterLongitude);
+        console.log(this.mapCenterZoomLevel);
+
         _map = L.map('leaflet-map', {
           preferCanvas: true,
         }).setView([
-          process.env.VUE_APP_VISU_MAP_CENTER_LATITUDE,
-          process.env.VUE_APP_VISU_MAP_CENTER_LONGITUDE
-        ], process.env.VUE_APP_VISU_MAP_ZOOM_LEVEL);
+          this.mapCenterLatitude,
+          this.mapCenterLongitude
+        ], this.mapCenterZoomLevel);
 
         _tileLayers.push(L.tileLayer(this.mapBaseLayerUrl, {
           attribution: `&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors`,
@@ -322,7 +358,7 @@
         console.log(`Loading dataset from ${this.dataSourceUrl}`);
         _data = await this.loadData(this.dataSourceUrl);
 
-        while(!_data.reverse().some(l => l.date === this.dateFilter)) {
+        while (!_data.reverse().some(l => l.date === this.dateFilter)) {
           const previousDay = new Date(this.dateFilter);
           previousDay.setDate(previousDay.getDate() - 1);
           const previousDayString = previousDay.toISOString().split('T')[0];