+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - If no data is available for the current day, fallback to last date with data
+- If the user location is known, center the map on his position with extra zoom : see new VUE_APP_VISU_MAP_ZOOM_LEVEL_LOCAL
 
 ## [1.1.2] 2020-03-31